- Refactor all of it with two processes and HTTP between CoAP and Flask
