+++ src/client.py
@@ -1,5 +1,6 @@
 import asyncio
 
+import aiocoap.resource as resource
 from aiocoap import *
 
 
@@ -15,9 +16,11 @@ def send_msg(host, path, payload):
     else:
         print("lol", response.code, response.payload)
 
-
 asyncio.get_event_loop().run_until_complete(send_msg('localhost', '03/i_am_alive', '0'))
-
 asyncio.get_event_loop().run_until_complete(send_msg('localhost', '03/button', '1000'))
-
 asyncio.get_event_loop().run_until_complete(send_msg('localhost', '03/button', '0001'))
+
+#root = resource.Site()
+#root.add_resource(('led',), BlockResource())
+#asyncio.async(Context.create_server_context(root))
+#asyncio.get_event_loop().run_forever()
